pyproject.toml
[project]
name = "ridger"
version = "0.1.0"
description = "Create ridge images"
authors = [
    {name = "Jacopo Farina",email = "jacopo1.farina@gmail.com"},
]
dependencies = [
    "Pillow>=10.1.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.1.6",
]
